fix after CR#2
